requirements.txt
Flask==3.0.0
mysql-connector-python==8.3.0
PyCryptodome==3.20.0
opencv-python==4.10.0.84
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.26.4
Pillow==10.2.0
